update view home, navs bar and create logic log out
