+++ app/page.tsx
@@ -1,4 +1,4 @@
-import { h1 } from "framer-motion/client";
+
 import LoginComponent from "./login/page";
 
 function App() {
