+++ app/register/page.tsx
@@ -2,13 +2,15 @@
 
 import React, { useState} from "react"; // Importa React y el hook useState
 import { useRouter } from "next/navigation"; // Importa el hook useRouter para la navegación programática
-import { User, Lock } from "lucide-react";
+import { User, Lock, Mail } from "lucide-react";
 import Link from "next/link";
 import Logoregister from "@/components/types/icons/logo-register";
+import Nav from "../navInitial/page"
 
 const RegisterComponent: React.FC = () => {
   const [email, setEmail] = useState(""); // Estado para almacenar el email
   const [password, setPassword] = useState(""); // Estado para almacenar la contraseña
+  const [name, setName] = useState("")
   const router = useRouter(); // Inicializa el router para la navegación
 
 
@@ -24,11 +26,12 @@ const RegisterComponent: React.FC = () => {
         headers: {
           "Content-Type": "application/json", // Indica el tipo de contenido
         },
-        body: JSON.stringify({ email, password }), // Convierte el objeto a JSON
+        body: JSON.stringify({ email, password, name }), // Convierte el objeto a JSON
       });
       
       setEmail(""); // Limpia el campo de email
       setPassword(""); // Limpia el campo de contraseña
+      setName("");
 
       alert("Cuenta creada")
       router.push('/login')
@@ -37,12 +40,27 @@ const RegisterComponent: React.FC = () => {
 };
 
   return (
+    <>
+    <Nav/>
     <div className="login-container"> 
     <div className="login-form"> 
       <h2 className="app-name">MerkaSavvy</h2> 
       <p className="welcome-back">Welcome !!!</p> 
       <h1 className="sign-up">Sign up</h1> 
       <form id="animation" onSubmit={handleSubmit}> 
+      <div className="input-group"> 
+          <label htmlFor="email">Name</label> 
+          <div className="input-with-icon"> 
+            <input
+              type="text"
+              id="name"
+              placeholder="Jon Doe"
+              value={name} // Valor del input vinculado al estado email
+              onChange={(e) => setName(e.target.value)} // Actualiza el estado email al cambiar
+            />
+            <User id="name" className="icon" size={20} /> 
+          </div>
+        </div>
         <div className="input-group"> 
           <label htmlFor="email">Email</label> 
           <div className="input-with-icon"> 
@@ -53,7 +71,7 @@ const RegisterComponent: React.FC = () => {
               value={email} // Valor del input vinculado al estado email
               onChange={(e) => setEmail(e.target.value)} // Actualiza el estado email al cambiar
             />
-            <User id="user" className="icon" size={20} /> 
+            <Mail id="user" className="icon" size={20} /> 
           </div>
         </div>
         <div className="input-group"> 
@@ -73,7 +91,7 @@ const RegisterComponent: React.FC = () => {
           SIGN UP →
         </button>
       </form>
-      <Link className="sign-in-link" href="register"> 
+      <Link className="sign-in-link" href="login"> 
         I have an account? Sign In
       </Link>
     </div>
@@ -83,6 +101,7 @@ const RegisterComponent: React.FC = () => {
       </div>
     </div>
   </div>
+  </>
   );
 };
