+++ db.json
@@ -74,6 +74,24 @@
       "id": "0082",
       "email": "ju4nc4r1d4d@gmail.com",
       "password": "1"
+    },
+    {
+      "id": "a117",
+      "email": "kevin.rodas.cuartas@gmail.com",
+      "password": "123456",
+      "name": "Kevin Rodas Cuartas"
+    },
+    {
+      "id": "746a",
+      "email": "edwin@gmail.com",
+      "password": "123456",
+      "name": "Edwin"
+    },
+    {
+      "id": "e49b",
+      "email": "ramirito@gmail.com",
+      "password": "1",
+      "name": "ramiro"
     }
   ],
   "productos": [
